[r00017.1] Update audio.git : 4e4fed1
git-reposnap-release_tag: r00017.1
git-reposnap-name: platform/hardware/qcom/audio
git-reposnap-prefix: hardware/qcom/audio
git-reposnap-fetch_url: https://source.codeaurora.org//platform/hardware/qcom/audio.git
git-reposnap-upstream: audio-userspace.lnx.2.2.r15-rel
git-reposnap-rev: 4e4fed1fc856763d9f0123b14dda3ee67eecd4e6

commit 4e4fed1fc856763d9f0123b14dda3ee67eecd4e6
Merge: 3644690 7d32a7f
Author: Linux Build Service Account <lnxbuild@localhost>
Date:   Sun Jan 28 21:50:10 2018 -0800

    Merge 7d32a7f3f58f3c12de445ede183fb5e137067bf5 on remote branch

+++ hardware/qcom/audio/hal/voice_extn/compress_voip.c
@@ -128,8 +128,8 @@ static int voip_set_volume(struct audio_device *adev, int volume)
     struct mixer_ctl *ctl;
     const char *mixer_ctl_name = "Voip Rx Gain";
     int vol_index = 0;
-    uint32_t set_values[ ] = {0,
-                              DEFAULT_VOLUME_RAMP_DURATION_MS};
+    long set_values[ ] = {0,
+                          DEFAULT_VOLUME_RAMP_DURATION_MS};
 
     ALOGV("%s: enter", __func__);
 
@@ -157,8 +157,8 @@ static int voip_set_mic_mute(struct audio_device *adev, bool state)
 {
     struct mixer_ctl *ctl;
     const char *mixer_ctl_name = "Voip Tx Mute";
-    uint32_t set_values[ ] = {0,
-                              DEFAULT_VOLUME_RAMP_DURATION_MS};
+    long set_values[ ] = {0,
+                          DEFAULT_VOLUME_RAMP_DURATION_MS};
 
     ALOGV("%s: enter, state=%d", __func__, state);
